+++ foo_ui_columns/ng_playlist/ng_playlist_style.cpp
@@ -3,7 +3,38 @@
 #include "ng_playlist.h"
 #include "ng_playlist_style.h"
 
+#include "tf_utils.h"
+
 namespace cui::panels::playlist_view {
+
+namespace {
+
+Colour get_colour_param(titleformat_hook_function_params& params, size_t index)
+{
+    const auto text = tf::get_param(params, index);
+    const auto colour_text = text.size() > 0 && text[0] == 3 ? text.substr(1) : text;
+    return Colour(mmh::strtoul_n(colour_text.data(), colour_text.size(), 0x10));
+}
+
+std::tuple<Colour, Colour> get_colour_pair_param(titleformat_hook_function_params& params, size_t index)
+{
+    const auto text = tf::get_param(params, index);
+    const auto colour_text = text.size() > 0 && text[0] == 3 ? text.substr(1) : text;
+    const auto colour_1 = mmh::strtoul_n(colour_text.data(), colour_text.size(), 0x10);
+
+    const auto colour_2 = [&] {
+        if (const auto bar_pos = colour_text.find_first_of('|'); bar_pos != std::string_view::npos) {
+            const auto colour_2_text = colour_text.substr(bar_pos + 1);
+            return mmh::strtoul_n(colour_2_text.data(), colour_2_text.size(), 0x10);
+        }
+        return 0xffffff - colour_1;
+    }();
+
+    return std::make_tuple(Colour(colour_1), Colour(colour_2));
+}
+
+} // namespace
+
 namespace style_cache_manager {
 
 pfc::list_t<SharedCellStyleData*> m_objects;
@@ -48,8 +79,8 @@ bool StyleTitleformatHook::process_field(
 {
     p_found_flag = false;
     {
-        if (p_name_length > 1 && !stricmp_utf8_ex(p_name, 1, "_", pfc_infinite)) {
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "text", pfc_infinite)) {
+        if (p_name_length > 1 && !stricmp_utf8_ex(p_name, 1, "_", SIZE_MAX)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "text", SIZE_MAX)) {
                 if (!text.get_ptr()) {
                     text.set_size(33);
                     text.fill(0);
@@ -59,7 +90,7 @@ bool StyleTitleformatHook::process_field(
                 p_found_flag = true;
                 return true;
             }
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "selected_text", pfc_infinite)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "selected_text", SIZE_MAX)) {
                 if (!selected_text.get_ptr()) {
                     selected_text.set_size(33);
                     selected_text.fill(0);
@@ -70,7 +101,7 @@ bool StyleTitleformatHook::process_field(
                 p_found_flag = true;
                 return true;
             }
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "back", pfc_infinite)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "back", SIZE_MAX)) {
                 if (!back.get_ptr()) {
                     back.set_size(33);
                     back.fill(0);
@@ -80,7 +111,7 @@ bool StyleTitleformatHook::process_field(
                 p_found_flag = true;
                 return true;
             }
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "selected_back", pfc_infinite)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "selected_back", SIZE_MAX)) {
                 if (!selected_back.get_ptr()) {
                     selected_back.set_size(33);
                     selected_back.fill(0);
@@ -91,7 +122,7 @@ bool StyleTitleformatHook::process_field(
                 p_found_flag = true;
                 return true;
             }
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "selected_back_no_focus", pfc_infinite)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "selected_back_no_focus", SIZE_MAX)) {
                 if (!selected_back_no_focus.get_ptr()) {
                     selected_back_no_focus.set_size(33);
                     selected_back_no_focus.fill(0);
@@ -103,7 +134,7 @@ bool StyleTitleformatHook::process_field(
                 p_found_flag = true;
                 return true;
             }
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "selected_text_no_focus", pfc_infinite)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "selected_text_no_focus", SIZE_MAX)) {
                 if (!selected_text_no_focus.get_ptr()) {
                     selected_text_no_focus.set_size(33);
                     selected_text_no_focus.fill(0);
@@ -115,7 +146,7 @@ bool StyleTitleformatHook::process_field(
                 p_found_flag = true;
                 return true;
             }
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "themed", pfc_infinite)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "themed", SIZE_MAX)) {
                 colours::helper p_helper(ColoursClient::id);
                 if (p_helper.get_themed()) {
                     p_out->write(titleformat_inputtypes::unknown, "1", 1);
@@ -123,7 +154,7 @@ bool StyleTitleformatHook::process_field(
                 }
                 return true;
             }
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "display_index", pfc_infinite)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "display_index", SIZE_MAX)) {
                 if (!m_index_text) {
                     m_index_text = fmt::format("{}", m_index + 1);
                 }
@@ -131,7 +162,7 @@ bool StyleTitleformatHook::process_field(
                 p_found_flag = true;
                 return true;
             }
-            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "is_group", pfc_infinite)) {
+            if (!stricmp_utf8_ex(p_name + 1, p_name_length - 1, "is_group", SIZE_MAX)) {
                 if (m_is_group) {
                     p_out->write(titleformat_inputtypes::unknown, "1", 1);
                     p_found_flag = true;
@@ -148,170 +179,75 @@ bool StyleTitleformatHook::process_function(titleformat_text_out* p_out, const c
     titleformat_hook_function_params* p_params, bool& p_found_flag)
 {
     p_found_flag = false;
-    if (!stricmp_utf8_ex(p_name, p_name_length, "set_style", pfc_infinite)) {
+    if (!stricmp_utf8_ex(p_name, p_name_length, "set_style", SIZE_MAX)) {
         if (p_params->get_param_count() >= 2) {
-            const char* name;
-            size_t name_length;
-            p_params->get_param(0, name, name_length);
-            if (!stricmp_utf8_ex(name, name_length, "text", pfc_infinite)) {
-                {
-                    const char* value;
-                    size_t value_length;
-                    p_params->get_param(1, value, value_length);
-                    if (value_length && *value == 3) {
-                        value++;
-                        value_length--;
-                    }
-                    p_colours.text_colour.set(mmh::strtoul_n(value, value_length, 0x10));
-                    if (value_length == 6 * 2 + 2 && value[6] == '|') {
-                        value += 7;
-                        value_length -= 7;
-                        p_colours.selected_text_colour.set(mmh::strtoul_n(value, value_length, 0x10));
-                    } else
-                        p_colours.selected_text_colour.set(0xffffff - p_colours.text_colour);
+            const auto name = tf::get_param(*p_params, 0);
+            if (!stricmp_utf8_ex(name.data(), name.size(), "text", SIZE_MAX)) {
+                if (p_params->get_param_count() == 2) {
+                    std::tie(p_colours.text_colour, p_colours.selected_text_colour)
+                        = get_colour_pair_param(*p_params, 1);
+                } else {
+                    p_colours.text_colour = get_colour_param(*p_params, 1);
+                    p_colours.selected_text_colour = get_colour_param(*p_params, 2);
                 }
-                if (p_params->get_param_count() >= 3) {
-                    {
-                        const char* value;
-                        size_t value_length;
-                        p_params->get_param(2, value, value_length);
-                        if (value_length && *value == 3) {
-                            value++;
-                            value_length--;
-                        }
-                        p_colours.selected_text_colour.set(mmh::strtoul_n(value, value_length, 0x10));
-                    }
-                }
-                if (p_params->get_param_count() >= 4) {
-                    const char* value;
-                    size_t value_length;
-                    p_params->get_param(3, value, value_length);
-                    if (value_length && *value == 3) {
-                        value++;
-                        value_length--;
-                    }
-                    p_colours.selected_text_colour_non_focus.set(mmh::strtoul_n(value, value_length, 0x10));
-                } else
-                    p_colours.selected_text_colour_non_focus.set(p_colours.selected_text_colour);
-            } else if (!stricmp_utf8_ex(name, name_length, "back", pfc_infinite)) {
-                {
-                    const char* value;
-                    size_t value_length;
-                    p_params->get_param(1, value, value_length);
-                    if (value_length && *value == 3) {
-                        value++;
-                        value_length--;
-                    }
-                    p_colours.background_colour.set(mmh::strtoul_n(value, value_length, 0x10));
 
-                    if (value_length == 6 * 2 + 2 && value[6] == '|') {
-                        value += 7;
-                        value_length -= 7;
-                        p_colours.selected_background_colour.set(mmh::strtoul_n(value, value_length, 0x10));
-                    } else
-                        p_colours.selected_background_colour.set(0xffffff - p_colours.background_colour);
+                if (p_params->get_param_count() >= 4)
+                    p_colours.selected_text_colour_non_focus = get_colour_param(*p_params, 3);
+                else
+                    p_colours.selected_text_colour_non_focus = p_colours.selected_text_colour;
+            } else if (!stricmp_utf8_ex(name.data(), name.size(), "back", SIZE_MAX)) {
+                if (p_params->get_param_count() == 2) {
+                    std::tie(p_colours.background_colour, p_colours.selected_background_colour)
+                        = get_colour_pair_param(*p_params, 1);
+                } else {
+                    p_colours.background_colour = get_colour_param(*p_params, 1);
+                    p_colours.selected_background_colour = get_colour_param(*p_params, 2);
                 }
-                if (p_params->get_param_count() >= 3) {
-                    {
-                        const char* value;
-                        size_t value_length;
-                        p_params->get_param(2, value, value_length);
-                        if (value_length && *value == 3) {
-                            value++;
-                            value_length--;
-                        }
-                        p_colours.selected_background_colour.set(mmh::strtoul_n(value, value_length, 0x10));
-                    }
-                    if (p_params->get_param_count() >= 4) {
-                        const char* value;
-                        size_t value_length;
-                        p_params->get_param(3, value, value_length);
-                        if (value_length && *value == 3) {
-                            value++;
-                            value_length--;
-                        }
-                        p_colours.selected_background_colour_non_focus.set(mmh::strtoul_n(value, value_length, 0x10));
-                    } else
-                        p_colours.selected_background_colour_non_focus.set(p_colours.selected_background_colour);
-                } else
-                    p_colours.selected_background_colour_non_focus.set(p_colours.selected_background_colour);
-            } else if (name_length >= 6 && !stricmp_utf8_ex(name, 6, "frame-", pfc_infinite)) {
-                const char* p_side = name;
-                p_side += 6;
-                if (!stricmp_utf8_ex(p_side, name_length - 6, "left", pfc_infinite)) {
-                    const char* value;
-                    size_t value_length;
-                    p_params->get_param(1, value, value_length);
-                    p_colours.use_frame_left = (value_length == 1 && *value == '1');
-                    if (p_params->get_param_count() >= 3) {
-                        {
-                            const char* value;
-                            size_t value_length;
-                            p_params->get_param(2, value, value_length);
-                            if (value_length && *value == 3) {
-                                value++;
-                                value_length--;
-                            }
-                            p_colours.frame_left.set(mmh::strtoul_n(value, value_length, 0x10));
-                        }
-                    }
-                } else if (!stricmp_utf8_ex(p_side, name_length - 6, "top", pfc_infinite)) {
-                    const char* value;
-                    size_t value_length;
-                    p_params->get_param(1, value, value_length);
-                    p_colours.use_frame_top = (value_length == 1 && *value == '1');
-                    if (p_params->get_param_count() >= 3) {
-                        {
-                            const char* value;
-                            size_t value_length;
-                            p_params->get_param(2, value, value_length);
-                            if (value_length && *value == 3) {
-                                value++;
-                                value_length--;
-                            }
-                            p_colours.frame_top.set(mmh::strtoul_n(value, value_length, 0x10));
-                        }
-                    }
-                } else if (!stricmp_utf8_ex(p_side, name_length - 6, "right", pfc_infinite)) {
-                    const char* value;
-                    size_t value_length;
-                    p_params->get_param(1, value, value_length);
-                    p_colours.use_frame_right = (value_length == 1 && *value == '1');
-                    if (p_params->get_param_count() >= 3) {
-                        {
-                            const char* value;
-                            size_t value_length;
-                            p_params->get_param(2, value, value_length);
-                            if (value_length && *value == 3) {
-                                value++;
-                                value_length--;
-                            }
-                            p_colours.frame_right.set(mmh::strtoul_n(value, value_length, 0x10));
-                        }
-                    }
-                } else if (!stricmp_utf8_ex(p_side, name_length - 6, "bottom", pfc_infinite)) {
-                    const char* value;
-                    size_t value_length;
-                    p_params->get_param(1, value, value_length);
-                    p_colours.use_frame_bottom = (value_length == 1 && *value == '1');
-                    if (p_params->get_param_count() >= 3) {
-                        {
-                            const char* value;
-                            size_t value_length;
-                            p_params->get_param(2, value, value_length);
-                            if (value_length && *value == 3) {
-                                value++;
-                                value_length--;
-                            }
-                            p_colours.frame_bottom.set(mmh::strtoul_n(value, value_length, 0x10));
-                        }
-                    }
+
+                if (p_params->get_param_count() >= 4)
+                    p_colours.selected_background_colour_non_focus = get_colour_param(*p_params, 3);
+                else
+                    p_colours.selected_background_colour_non_focus = p_colours.selected_background_colour;
+            } else if (!stricmp_utf8_ex(name.data(), name.size(), "group-line", SIZE_MAX)) {
+                const auto param_1 = tf::get_param(*p_params, 1);
+                p_colours.show_group_line = param_1 == "1"sv || param_1 == "true"sv;
+
+                if (p_params->get_param_count() == 3)
+                    p_colours.group_line_colour = get_colour_param(*p_params, 2);
+            } else if (name.size() >= 6 && !stricmp_utf8_ex(name.data(), 6, "frame-", SIZE_MAX)) {
+                const auto side_name = name.substr(6);
+                const auto param_1 = tf::get_param(*p_params, 1);
+                const auto use_frame = param_1 == "1"sv || param_1 == "true"sv;
+                const auto colour = p_params->get_param_count() >= 3
+                    ? std::make_optional(get_colour_param(*p_params, 2))
+                    : std::nullopt;
+
+                if (!stricmp_utf8_ex(side_name.data(), side_name.size(), "left", SIZE_MAX)) {
+                    p_colours.use_frame_left = use_frame;
+
+                    if (colour)
+                        p_colours.frame_left = *colour;
+                } else if (!stricmp_utf8_ex(side_name.data(), side_name.size(), "top", SIZE_MAX)) {
+                    p_colours.use_frame_top = use_frame;
+
+                    if (colour)
+                        p_colours.frame_top = *colour;
+                } else if (!stricmp_utf8_ex(side_name.data(), side_name.size(), "right", SIZE_MAX)) {
+                    p_colours.use_frame_right = use_frame;
+
+                    if (colour)
+                        p_colours.frame_right = *colour;
+                } else if (!stricmp_utf8_ex(side_name.data(), side_name.size(), "bottom", SIZE_MAX)) {
+                    p_colours.use_frame_bottom = use_frame;
+
+                    if (colour)
+                        p_colours.frame_bottom = *colour;
                 }
             }
             p_found_flag = true;
             return true;
         }
-    } else if (!stricmp_utf8_ex(p_name, p_name_length, "calculate_blend_target", pfc_infinite)) {
+    } else if (!stricmp_utf8_ex(p_name, p_name_length, "calculate_blend_target", SIZE_MAX)) {
         if (p_params->get_param_count() == 1) {
             const char* p_val;
             size_t p_val_length;
@@ -328,7 +264,7 @@ bool StyleTitleformatHook::process_function(titleformat_text_out* p_out, const c
             p_found_flag = true;
             return true;
         }
-    } else if (!stricmp_utf8_ex(p_name, p_name_length, "offset_colour", pfc_infinite)) {
+    } else if (!stricmp_utf8_ex(p_name, p_name_length, "offset_colour", SIZE_MAX)) {
         if (p_params->get_param_count() == 3) {
             const char* p_val;
             const char* p_val2;